move tree to its own component in the components folder and finish primary functionality
